Move input UI into own module
