+++ packages/cli-kit/src/private/node/session.ts
@@ -21,7 +21,7 @@ import {AbortError, BugError} from '../../public/node/error.js'
 import {partnersRequest} from '../../public/node/api/partners.js'
 import {normalizeStoreFqdn, partnersFqdn, identityFqdn} from '../../public/node/context/fqdn.js'
 import {openURL} from '../../public/node/system.js'
-import {keypress} from '../../public/node/ui.js'
+import {keypress} from '../../public/node/ui-inputs.js'
 import {getIdentityTokenInformation, getPartnersToken} from '../../public/node/environment.js'
 import {gql} from 'graphql-request'
 import {AdminSession} from '@shopify/cli-kit/node/session'
