+++ packages/cli-kit/src/private/node/session/authorize.test.ts
@@ -5,7 +5,7 @@ import {randomHex, base64URLEncode} from '../../../public/node/crypto.js'
 import {openURL} from '../../../public/node/system.js'
 import {identityFqdn} from '../../../public/node/context/fqdn.js'
 import {CancelExecution} from '../../../public/node/error.js'
-import {renderConfirmationPrompt} from '../../../public/node/ui.js'
+import {renderConfirmationPrompt} from '../../../public/node/ui-inputs.js'
 import {checkPort} from 'get-port-please'
 import {killPortProcess} from 'kill-port-process'
 
@@ -14,7 +14,7 @@ import {describe, test, expect, vi} from 'vitest'
 vi.mock('../../../public/node/system.js')
 vi.mock('./redirect-listener')
 vi.mock('../../../public/node/crypto.js')
-vi.mock('../../../public/node/ui.js')
+vi.mock('../../../public/node/ui-inputs.js')
 vi.mock('../../../public/node/context/fqdn.js')
 vi.mock('./identity')
 vi.mock('get-port-please')
