+++ packages/cli-kit/src/private/node/session/device-authorization.test.ts
@@ -8,7 +8,7 @@ import {IdentityToken} from './schema.js'
 import {exchangeDeviceCodeForAccessToken} from './exchange.js'
 import {identityFqdn} from '../../../public/node/context/fqdn.js'
 import {shopifyFetch} from '../../../public/node/http.js'
-import {isTTY} from '../../../public/node/ui.js'
+import {isTTY} from '../../../public/node/ui-inputs.js'
 import {err, ok} from '../../../public/node/result.js'
 import {beforeEach, describe, expect, test, vi} from 'vitest'
 import {Response} from 'node-fetch'
@@ -16,7 +16,7 @@ import {Response} from 'node-fetch'
 vi.mock('../../../public/node/context/fqdn.js')
 vi.mock('./identity')
 vi.mock('../../../public/node/http.js')
-vi.mock('../../../public/node/ui.js')
+vi.mock('../../../public/node/ui-inputs.js')
 vi.mock('./exchange.js')
 
 beforeEach(() => {
