+++ packages/cli-kit/src/public/node/ui.tsx
@@ -1,20 +1,8 @@
 /* eslint-disable tsdoc/syntax */
-import {AbortError, AbortSilentError, FatalError as Fatal, FatalErrorType} from './error.js'
-import {
-  collectLog,
-  consoleError,
-  consoleLog,
-  Logger,
-  LogLevel,
-  outputContent,
-  outputDebug,
-  outputToken,
-  outputWhereAppropriate,
-} from './output.js'
+import {FatalError as Fatal, FatalErrorType} from './error.js'
+import {collectLog, consoleError, consoleLog, Logger, LogLevel, outputWhereAppropriate} from './output.js'
 import {isUnitTest} from './context/local.js'
-import {terminalSupportsRawMode} from './system.js'
 import {AbortController} from './abort.js'
-import {runWithTimer} from './metadata.js'
 import {ConcurrentOutput, ConcurrentOutputProps} from '../../private/node/ui/components/ConcurrentOutput.js'
 import {handleCtrlC, render, renderOnce} from '../../private/node/ui.js'
 import {alert, AlertOptions} from '../../private/node/ui/alert.js'
@@ -22,37 +10,15 @@ import {CustomSection} from '../../private/node/ui/components/Alert.js'
 import {FatalError} from '../../private/node/ui/components/FatalError.js'
 import ScalarDict from '../../private/node/ui/components/Table/ScalarDict.js'
 import {Table, TableColumn, TableProps} from '../../private/node/ui/components/Table/Table.js'
-import {
-  Token,
-  tokenItemToString,
-  InlineToken,
-  LinkToken,
-  ListToken,
-  TokenItem,
-} from '../../private/node/ui/components/TokenizedText.js'
-import {
-  DangerousConfirmationPrompt,
-  DangerousConfirmationPromptProps,
-} from '../../private/node/ui/components/DangerousConfirmationPrompt.js'
-import {SelectPrompt, SelectPromptProps} from '../../private/node/ui/components/SelectPrompt.js'
+import {Token, InlineToken, LinkToken, ListToken, TokenItem} from '../../private/node/ui/components/TokenizedText.js'
+
 import {Tasks, Task} from '../../private/node/ui/components/Tasks.js'
-import {TextPrompt, TextPromptProps} from '../../private/node/ui/components/TextPrompt.js'
-import {AutocompletePromptProps, AutocompletePrompt} from '../../private/node/ui/components/AutocompletePrompt.js'
 import {InfoTableSection} from '../../private/node/ui/components/Prompts/InfoTable.js'
 import {recordUIEvent, resetRecordedSleep} from '../../private/node/demo-recorder.js'
 import {InfoMessageProps} from '../../private/node/ui/components/Prompts/InfoMessage.js'
+import {PartialBy} from '../common/ts/partial-by.js'
 import React from 'react'
 import {Key as InkKey, RenderOptions} from 'ink'
-
-type PartialBy<T, TKey extends keyof T> = Omit<T, TKey> & Partial<Pick<T, TKey>>
-
-interface UIDebugOptions {
-  /** If true, don't check if the current terminal is interactive or not */
-  skipTTYCheck?: boolean
-}
-const defaultUIDebugOptions: UIDebugOptions = {
-  skipTTYCheck: false,
-}
 
 export interface RenderConcurrentOptions extends PartialBy<ConcurrentOutputProps, 'abortSignal'> {
   renderOptions?: RenderOptions
@@ -255,229 +221,6 @@ export function renderFatalError(error: Fatal, {renderOptions}: RenderFatalError
   return renderOnce(<FatalError error={error} />, {logLevel: 'error', logger: consoleError, renderOptions})
 }
 
-export interface RenderSelectPromptOptions<T> extends Omit<SelectPromptProps<T>, 'onSubmit'> {
-  isConfirmationPrompt?: boolean
-  renderOptions?: RenderOptions
-}
-
-/**
- * Renders a select prompt to the console.
- * @example
- * ?  Associate your project with the org Castile Ventures?
- *
- *    ┃  Add
- *    ┃  • new-ext
- *    ┃
- *    ┃  Remove
- *    ┃  • integrated-demand-ext
- *    ┃  • order-discount
- *
- *    Automations
- *    >  fifth
- *       sixth
- *
- *    Merchant Admin
- *       eighth
- *       ninth
- *
- *    Other
- *       first
- *       second
- *       third (limit reached)
- *       fourth
- *       seventh
- *       tenth
- *
- *    Press ↑↓ arrows to select, enter to confirm.
- *
- */
-// eslint-disable-next-line max-params
-export async function renderSelectPrompt<T>(
-  {renderOptions, isConfirmationPrompt, ...props}: RenderSelectPromptOptions<T>,
-  uiDebugOptions: UIDebugOptions = defaultUIDebugOptions,
-): Promise<T> {
-  throwInNonTTY({message: props.message, stdin: renderOptions?.stdin}, uiDebugOptions)
-
-  if (!isConfirmationPrompt) {
-    recordUIEvent({type: 'selectPrompt', properties: {renderOptions, ...props}})
-  }
-
-  return runWithTimer('cmd_all_timing_prompts_ms')(async () => {
-    let selectedValue: T
-    try {
-      await render(
-        <SelectPrompt
-          {...props}
-          onSubmit={(value: T) => {
-            selectedValue = value
-          }}
-        />,
-        {
-          ...renderOptions,
-          exitOnCtrlC: false,
-        },
-      )
-      return selectedValue!
-    } finally {
-      resetRecordedSleep()
-    }
-  })
-}
-
-export interface RenderConfirmationPromptOptions
-  extends Pick<SelectPromptProps<boolean>, 'message' | 'infoTable' | 'infoMessage' | 'abortSignal'> {
-  confirmationMessage?: string
-  cancellationMessage?: string
-  renderOptions?: RenderOptions
-  defaultValue?: boolean
-}
-
-/**
- * Renders a confirmation prompt to the console.
- * @example
- * ?  Delete the following themes from the store?
- *
- *    ┃  Info message title
- *    ┃
- *    ┃  Info message body
- *    ┃
- *    ┃  • first theme (#1)
- *    ┃  • second theme (#2)
- *
- * >  (y) Yes, confirm changes
- *    (n) Cancel
- *
- *    Press ↑↓ arrows to select, enter or a shortcut to
- *    confirm.
- *
- */
-export async function renderConfirmationPrompt({
-  message,
-  infoTable,
-  confirmationMessage = 'Yes, confirm',
-  cancellationMessage = 'No, cancel',
-  renderOptions,
-  defaultValue = true,
-  abortSignal,
-  infoMessage,
-}: RenderConfirmationPromptOptions): Promise<boolean> {
-  // eslint-disable-next-line prefer-rest-params
-  recordUIEvent({type: 'confirmationPrompt', properties: arguments[0]})
-
-  const choices = [
-    {
-      label: confirmationMessage,
-      value: true,
-      key: 'y',
-    },
-    {
-      label: cancellationMessage,
-      value: false,
-      key: 'n',
-    },
-  ]
-
-  return renderSelectPrompt({
-    choices,
-    message,
-    infoTable,
-    renderOptions,
-    defaultValue,
-    isConfirmationPrompt: true,
-    abortSignal,
-    infoMessage,
-  })
-}
-
-export interface RenderAutocompleteOptions<T>
-  extends PartialBy<Omit<AutocompletePromptProps<T>, 'onSubmit'>, 'search'> {
-  renderOptions?: RenderOptions
-}
-
-/**
- * Renders an autocomplete prompt to the console.
- * @example
- * ?  Select a template:   Type to search...
- *
- *    ┃  Info message title
- *    ┃
- *    ┃  Info message body
- *
- * >  first
- *    second
- *    third
- *    fourth
- *    fifth
- *    sixth
- *    seventh
- *    eighth
- *    ninth
- *    tenth
- *    eleventh
- *    twelfth
- *    thirteenth
- *    fourteenth
- *    fifteenth
- *    sixteenth
- *    seventeenth
- *    eighteenth
- *    nineteenth (disabled)
- *    twentieth
- *    twenty-first
- *    twenty-second
- *    twenty-third
- *    twenty-fourth
- *    twenty-fifth
- *
- *    Press ↑↓ arrows to select, enter to confirm.
- *
- */
-// eslint-disable-next-line max-params
-export async function renderAutocompletePrompt<T>(
-  {renderOptions, ...props}: RenderAutocompleteOptions<T>,
-  uiDebugOptions: UIDebugOptions = defaultUIDebugOptions,
-): Promise<T> {
-  throwInNonTTY({message: props.message, stdin: renderOptions?.stdin}, uiDebugOptions)
-
-  // eslint-disable-next-line prefer-rest-params
-  recordUIEvent({type: 'autocompletePrompt', properties: arguments[0]})
-
-  const newProps = {
-    search(term: string) {
-      const lowerTerm = term.toLowerCase()
-      return Promise.resolve({
-        data: props.choices.filter((item) => {
-          return (
-            item.label.toLowerCase().includes(lowerTerm) || (item.group && item.group.toLowerCase().includes(lowerTerm))
-          )
-        }),
-      })
-    },
-    ...props,
-  }
-
-  return runWithTimer('cmd_all_timing_prompts_ms')(async () => {
-    let selectedValue: T
-    try {
-      await render(
-        <AutocompletePrompt
-          {...newProps}
-          onSubmit={(value: T) => {
-            selectedValue = value
-          }}
-        />,
-        {
-          ...renderOptions,
-          exitOnCtrlC: false,
-        },
-      )
-      return selectedValue!
-    } finally {
-      resetRecordedSleep()
-    }
-  })
-}
-
 interface RenderTableOptions<T extends ScalarDict> extends TableProps<T> {
   renderOptions?: RenderOptions
 }
@@ -532,109 +275,6 @@ export async function renderTasks<TContext>(tasks: Task<TContext>[], {renderOpti
   })
 }
 
-export interface RenderTextPromptOptions extends Omit<TextPromptProps, 'onSubmit'> {
-  renderOptions?: RenderOptions
-}
-
-/**
- * Renders a text prompt to the console.
- * @example
- * ?  App project name (can be changed later):
- * >  expansive commerce app
- *    ▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔
- *
- */
-// eslint-disable-next-line max-params
-export async function renderTextPrompt(
-  {renderOptions, ...props}: RenderTextPromptOptions,
-  uiDebugOptions: UIDebugOptions = defaultUIDebugOptions,
-): Promise<string> {
-  throwInNonTTY({message: props.message, stdin: renderOptions?.stdin}, uiDebugOptions)
-
-  // eslint-disable-next-line prefer-rest-params
-  recordUIEvent({type: 'textPrompt', properties: arguments[0]})
-
-  return runWithTimer('cmd_all_timing_prompts_ms')(async () => {
-    let enteredText = ''
-    try {
-      await render(
-        <TextPrompt
-          {...props}
-          onSubmit={(value: string) => {
-            enteredText = value
-          }}
-        />,
-        {
-          ...renderOptions,
-          exitOnCtrlC: false,
-        },
-      )
-      return enteredText
-    } finally {
-      resetRecordedSleep()
-    }
-  })
-}
-
-export interface RenderDangerousConfirmationPromptOptions extends Omit<DangerousConfirmationPromptProps, 'onSubmit'> {
-  renderOptions?: RenderOptions
-}
-
-/**
- * Renders a dangerous confirmation prompt to the console, forcing the user to
- * type a confirmation string to proceed.
- * @example
- * ?  Release a new version of nightly-app-2023-06-19?
- *
- *    ┃  Includes:
- *    ┃  + web-px (new)
- *    ┃  + sub-ui-ext
- *    ┃  + theme-app-ext
- *    ┃  + paymentify (from Partner Dashboard)
- *    ┃
- *    ┃  Removes:
- *    ┃  - prod-discount-fun
- *    ┃
- *    ┃  This can permanently delete app user data.
- *
- *    Type nightly-app-2023-06-19 to confirm, or press Escape
- *    to cancel.
- * >  █
- *    ▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔
- *
- */
-// eslint-disable-next-line max-params
-export async function renderDangerousConfirmationPrompt(
-  {renderOptions, ...props}: RenderDangerousConfirmationPromptOptions,
-  uiDebugOptions: UIDebugOptions = defaultUIDebugOptions,
-): Promise<boolean> {
-  throwInNonTTY({message: props.message, stdin: renderOptions?.stdin}, uiDebugOptions)
-
-  // eslint-disable-next-line prefer-rest-params
-  recordUIEvent({type: 'dangerousConfirmationPrompt', properties: arguments[0]})
-
-  return runWithTimer('cmd_all_timing_prompts_ms')(async () => {
-    let confirmed: boolean
-    try {
-      await render(
-        <DangerousConfirmationPrompt
-          {...props}
-          onSubmit={(value: boolean) => {
-            confirmed = value
-          }}
-        />,
-        {
-          ...renderOptions,
-          exitOnCtrlC: false,
-        },
-      )
-      return confirmed!
-    } finally {
-      resetRecordedSleep()
-    }
-  })
-}
-
 interface RenderTextOptions {
   text: string
   logLevel?: LogLevel
@@ -654,67 +294,6 @@ export function renderText({text, logLevel = 'info', logger = consoleLog}: Rende
   if (isUnitTest()) collectLog(logLevel, textWithLineReturn)
   outputWhereAppropriate(logLevel, logger, textWithLineReturn)
   return textWithLineReturn
-}
-
-/** Waits for any key to be pressed except Ctrl+C which will terminate the process. */
-// eslint-disable-next-line max-params
-export const keypress = async (stdin = process.stdin, uiDebugOptions: UIDebugOptions = defaultUIDebugOptions) => {
-  throwInNonTTY({message: 'Press any key'}, uiDebugOptions)
-
-  return runWithTimer('cmd_all_timing_prompts_ms')(() => {
-    // eslint-disable-next-line max-params
-    return new Promise((resolve, reject) => {
-      const handler = (buffer: Buffer) => {
-        stdin.setRawMode(false)
-
-        const bytes = Array.from(buffer)
-
-        if (bytes.length && bytes[0] === 3) {
-          outputDebug('Canceled keypress, User pressed CTRL+C')
-          reject(new AbortSilentError())
-        }
-        stdin.unref()
-        process.nextTick(resolve)
-      }
-
-      stdin.setRawMode(true)
-      stdin.once('data', handler)
-
-      // We want to indicate that we're still using stdin, so that the process
-      // doesn't exit early.
-      stdin.ref()
-    })
-  })
-}
-
-interface IsTTYOptions {
-  stdin?: NodeJS.ReadStream
-  uiDebugOptions?: UIDebugOptions
-}
-
-export function isTTY({stdin = undefined, uiDebugOptions = defaultUIDebugOptions}: IsTTYOptions = {}) {
-  return Boolean(uiDebugOptions.skipTTYCheck || stdin || terminalSupportsRawMode())
-}
-
-interface ThrowInNonTTYOptions {
-  message: TokenItem
-  stdin?: NodeJS.ReadStream
-}
-
-// eslint-disable-next-line max-params
-function throwInNonTTY({message, stdin = undefined}: ThrowInNonTTYOptions, uiDebugOptions: UIDebugOptions) {
-  if (isTTY({stdin, uiDebugOptions})) return
-
-  const promptText = tokenItemToString(message)
-  const errorMessage = `Failed to prompt:
-
-${outputContent`${outputToken.cyan(promptText)}`.value}
-
-This usually happens when running a command non-interactively, for example in a CI environment, or when piping input from another process.`
-  throw new AbortError(
-    errorMessage,
-    'To resolve this, specify the option in the command, or run the command in an interactive environment such as your local terminal.',
-  )
 }
 
 export type Key = InkKey
