+++ packages/plugin-did-you-mean/src/index.ts
@@ -4,7 +4,8 @@ import AutocorrectStatus from './commands/config/autocorrect/status.js'
 import AutocorrectOn from './commands/config/autocorrect/on.js'
 import {Hook} from '@oclif/core'
 import {bigram} from 'n-gram'
-import {renderConfirmationPrompt, renderFatalError, renderInfo} from '@shopify/cli-kit/node/ui'
+import {renderFatalError, renderInfo} from '@shopify/cli-kit/node/ui'
+import {renderConfirmationPrompt} from '@shopify/cli-kit/node/ui-inputs'
 import {AbortError} from '@shopify/cli-kit/node/error'
 import {outputDebug} from '@shopify/cli-kit/node/output'
