+++ packages/theme/src/cli/commands/theme/init.ts
@@ -4,7 +4,7 @@ import {cloneRepoAndCheckoutLatestTag, cloneRepo} from '../../services/init.js'
 import {Args, Flags} from '@oclif/core'
 import {globalFlags} from '@shopify/cli-kit/node/cli'
 import {generateRandomNameForSubdirectory} from '@shopify/cli-kit/node/fs'
-import {renderTextPrompt} from '@shopify/cli-kit/node/ui'
+import {renderTextPrompt} from '@shopify/cli-kit/node/ui-inputs'
 import {joinPath} from '@shopify/cli-kit/node/path'
 
 export default class Init extends ThemeCommand {
