+++ packages/theme/src/cli/services/publish.test.ts
@@ -1,12 +1,14 @@
 import {publish} from './publish.js'
 import {findOrSelectTheme} from '../utilities/theme-selector.js'
-import {renderSuccess, renderConfirmationPrompt} from '@shopify/cli-kit/node/ui'
+import {renderSuccess} from '@shopify/cli-kit/node/ui'
+import {renderConfirmationPrompt} from '@shopify/cli-kit/node/ui-inputs'
 import {test, describe, expect, vi} from 'vitest'
 import {Theme} from '@shopify/cli-kit/node/themes/types'
 import {publishTheme} from '@shopify/cli-kit/node/themes/api'
 
 vi.mock('@shopify/cli-kit/node/system')
 vi.mock('@shopify/cli-kit/node/ui')
+vi.mock('@shopify/cli-kit/node/ui-inputs')
 vi.mock('@shopify/cli-kit/node/themes/api')
 vi.mock('../utilities/theme-selector.js')
